+++ src/API/__init__.py
@@ -1,0 +1,5 @@
+"""New-style API imports
+"""
+
+from .Info import *
+from .Limits import *
